Use BigDecimal; remove presenca cascade
Replaced Double with BigDecimal for monetary fields (valor, valorMensalidade) in Matricula entity and related DTOs (Matricula, CadastroMatriculaDTO, AtualizarMatriculaDTO, MatriculaDTO), adding necessary imports and updating constructors/getters/setters. Updated SQL DDL (sql/ddl.sql.txt) to remove ON DELETE CASCADE from the presenca -> matricula foreign key to avoid cascading deletes of presenca records. Files changed: src/main/java/com/br/gerenciadoraulas/model/Matricula.java, src/main/java/com/br/gerenciadoraulas/dto/{CadastroMatriculaDTO,AtualizarMatriculaDTO,MatriculaDTO}. SQL: sql/ddl.sql.txt.

+++ src/main/java/com/br/gerenciadoraulas/dto/MatriculaDTO.java
@@ -4,6 +4,7 @@ import com.br.gerenciadoraulas.model.Aluno;
 import com.br.gerenciadoraulas.model.Matricula;
 import com.br.gerenciadoraulas.model.ProgramaAula;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class MatriculaDTO {
@@ -16,8 +17,8 @@ public class MatriculaDTO {
     private Boolean presente;
     private Long presencaId;
     private Boolean flAtivo;
-    private Double valor;
-    private Double valorMensalidade;
+    private BigDecimal valor;
+    private BigDecimal valorMensalidade;
     private Integer diaVencimento;
 
     public MatriculaDTO(){}
@@ -116,19 +117,19 @@ public class MatriculaDTO {
         this.flAtivo = flAtivo;
     }
 
-    public Double getValor() {
+    public BigDecimal getValor() {
         return valor;
     }
 
-    public void setValor(Double valor) {
+    public void setValor(BigDecimal valor) {
         this.valor = valor;
     }
 
-    public Double getValorMensalidade() {
+    public BigDecimal getValorMensalidade() {
         return valorMensalidade;
     }
 
-    public void setValorMensalidade(Double valorMensalidade) {
+    public void setValorMensalidade(BigDecimal valorMensalidade) {
         this.valorMensalidade = valorMensalidade;
     }
